// src/app/_models/scips.interface.ts
export interface IScips {  
    AV1: Boolean
    AV1_Degree: String
    AV1_Justification: String
    AV1_Value: String
    AV2: Boolean
    AV2_Degree: String
    AV2_Justification: String
    AV2_Value: String
    AV3: Boolean
    AV3_Degree: String
    AV3_Justification: String
    AV3_Value: String
    AV4: Boolean
    AV4_Degree: String
    AV4_Justification: String
    AV4_Value: String
    AV5: Boolean
    AV5_Degree: String
    AV5_Justification: String
    AV5_Value: String
    ActiveFlag: Boolean
    AlternateControlExample: String
    Analysis: String
    AppliesTo: String
    AssessmentStatus: String
    CommanControlExample: String
    ComplexityCategoryId: String
    ControlApplicability: Boolean
    ControlOrRequirementId: Number
    CorrectiveAction: String
    CorrectiveActionAnswerOptions: String
    Decision: String
    DecisionAnswerOptions: String
    Focus: String
    ImplementControl: Boolean
    Observation: String
    Purpose: String
    UseAlternate: Boolean
}